Moved burp extension modules to 'burp_extension/'
